# GREENALPHA/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GreenAlpha Carbon Footprint Engine",
    "description": "Advanced carbon footprint calculation and sustainability analytics API with <500ms response time",
    "version": "1.0.0"
  },
  "paths": {
    "/carbon/calculate": {
      "post": {
        "tags": [
          "Carbon Calculation",
          "Carbon Calculator"
        ],
        "summary": "Calculate Carbon Footprint",
        "description": "Calculate carbon footprint for a product shipment\n\nThis endpoint calculates the complete carbon footprint including:\n- Production emissions (Scope 1 & 2)\n- Transportation emissions (Scope 3)\n- Carbon cost estimation\n- Trading opportunities\n- Detailed breakdown and uncertainty analysis\n\nResponse time target: <500ms for 95% of requests",
        "operationId": "calculate_carbon_footprint_carbon_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarbonCalculationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarbonCalculationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/carbon/calculate/batch": {
      "post": {
        "tags": [
          "Carbon Calculation",
          "Carbon Calculator"
        ],
        "summary": "Calculate Batch Carbon Footprint",
        "description": "Calculate carbon footprint for multiple products in batch\n\nProcesses up to 100 calculations in parallel for optimal performance.\nReturns individual results plus optional batch summary.",
        "operationId": "calculate_batch_carbon_footprint_carbon_calculate_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCalculationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/recommendations/recommend/suppliers": {
      "post": {
        "tags": [
          "Supplier Recommendations"
        ],
        "summary": "Get Supplier Recommendations",
        "description": "Get supplier recommendations based on Multi-Criteria Decision Analysis (MCDA).",
        "operationId": "get_supplier_recommendations_recommendations_recommend_suppliers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRecommendationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendations/recommend/alternatives/{supplier_id}": {
      "get": {
        "tags": [
          "Supplier Recommendations"
        ],
        "summary": "Find Alternative Suppliers",
        "description": "Find alternative suppliers similar to a given supplier.",
        "operationId": "find_alternative_suppliers_recommendations_recommend_alternatives__supplier_id__get",
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          },
          {
            "name": "top_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 3,
              "title": "Top N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendations/recommend/batch": {
      "post": {
        "tags": [
          "Supplier Recommendations"
        ],
        "summary": "Placeholder for Batch Recommendations",
        "description": "A placeholder endpoint for future bulk recommendation processing.\n\nIn a real implementation, this would likely be an asynchronous task.",
        "operationId": "batch_recommendations_recommendations_recommend_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/RecommendationRequest"
                },
                "type": "array",
                "title": "Requests"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transport/optimize/route": {
      "post": {
        "tags": [
          "Transport Optimization"
        ],
        "summary": "Optimize Single Route",
        "description": "Get optimal transport routes for a single origin-destination pair",
        "operationId": "optimize_single_route_transport_optimize_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/arbitrage/arbitrage/comprehensive-report": {
      "post": {
        "tags": [
          "Carbon Arbitrage"
        ],
        "summary": "Generate Comprehensive Report",
        "description": "Generate comprehensive carbon arbitrage analysis report",
        "operationId": "generate_comprehensive_report_arbitrage_arbitrage_comprehensive_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArbitrageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArbitrageRequest": {
        "properties": {
          "countries": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Countries",
            "description": "Specific countries to analyze"
          },
          "min_volume_threshold": {
            "type": "number",
            "title": "Min Volume Threshold",
            "description": "Minimum volume threshold in tonnes",
            "default": 100000
          },
          "time_horizon_months": {
            "type": "integer",
            "title": "Time Horizon Months",
            "description": "Analysis time horizon in months",
            "default": 12
          }
        },
        "type": "object",
        "title": "ArbitrageRequest",
        "description": "Request model for arbitrage analysis"
      },
      "BatchCalculationRequest": {
        "properties": {
          "calculations": {
            "items": {
              "$ref": "#/components/schemas/CarbonCalculationRequest"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Calculations",
            "description": "List of calculations (max 100)"
          },
          "include_summary": {
            "type": "boolean",
            "title": "Include Summary",
            "description": "Include batch summary",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "calculations"
        ],
        "title": "BatchCalculationRequest",
        "description": "Batch carbon footprint calculation request"
      },
      "CarbonCalculationRequest": {
        "properties": {
          "product_name": {
            "type": "string",
            "title": "Product Name",
            "description": "Product name or category"
          },
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "Quantity of products",
            "default": 1.0
          },
          "origin_country": {
            "type": "string",
            "title": "Origin Country",
            "description": "Origin country code (3-letter)"
          },
          "destination_country": {
            "type": "string",
            "title": "Destination Country",
            "description": "Destination country code (3-letter)"
          },
          "transport_mode": {
            "type": "string",
            "title": "Transport Mode",
            "description": "Transportation mode",
            "default": "road_truck"
          },
          "product_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Category",
            "description": "Product category",
            "default": "general"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg",
            "description": "Total weight in kg"
          },
          "origin_latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Latitude"
          },
          "origin_longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Longitude"
          },
          "destination_latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Latitude"
          },
          "destination_longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Longitude"
          },
          "custom_emission_factors": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Emission Factors",
            "description": "Custom emission factors override"
          }
        },
        "type": "object",
        "required": [
          "product_name",
          "origin_country",
          "destination_country"
        ],
        "title": "CarbonCalculationRequest",
        "description": "Carbon footprint calculation request model"
      },
      "CarbonCalculationResponse": {
        "properties": {
          "total_emissions_kg_co2e": {
            "type": "number",
            "title": "Total Emissions Kg Co2E"
          },
          "production_emissions": {
            "type": "number",
            "title": "Production Emissions"
          },
          "transportation_emissions": {
            "type": "number",
            "title": "Transportation Emissions"
          },
          "scope_1_emissions": {
            "type": "number",
            "title": "Scope 1 Emissions"
          },
          "scope_2_emissions": {
            "type": "number",
            "title": "Scope 2 Emissions"
          },
          "scope_3_emissions": {
            "type": "number",
            "title": "Scope 3 Emissions"
          },
          "carbon_cost_usd": {
            "type": "number",
            "title": "Carbon Cost Usd"
          },
          "carbon_trading_opportunities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Carbon Trading Opportunities"
          },
          "calculation_confidence": {
            "type": "number",
            "title": "Calculation Confidence"
          },
          "response_time_ms": {
            "type": "number",
            "title": "Response Time Ms"
          },
          "calculation_method": {
            "type": "string",
            "title": "Calculation Method"
          },
          "data_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Sources"
          },
          "production_breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Production Breakdown"
          },
          "transport_breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Transport Breakdown"
          },
          "uncertainty_analysis": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Uncertainty Analysis"
          },
          "recommendations": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendations"
          }
        },
        "type": "object",
        "required": [
          "total_emissions_kg_co2e",
          "production_emissions",
          "transportation_emissions",
          "scope_1_emissions",
          "scope_2_emissions",
          "scope_3_emissions",
          "carbon_cost_usd",
          "carbon_trading_opportunities",
          "calculation_confidence",
          "response_time_ms",
          "calculation_method",
          "data_sources",
          "production_breakdown",
          "transport_breakdown",
          "uncertainty_analysis"
        ],
        "title": "CarbonCalculationResponse",
        "description": "Carbon footprint calculation response model"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedRecommendationResponse": {
        "properties": {
          "total_suppliers": {
            "type": "integer",
            "title": "Total Suppliers"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total_suppliers",
          "page",
          "page_size",
          "total_pages",
          "items"
        ],
        "title": "PaginatedRecommendationResponse",
        "description": "Paginated response model for recommendations."
      },
      "RecommendationRequest": {
        "properties": {
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Weights",
            "example": {
              "cost_index": 0.3,
              "esg_score": 0.4,
              "geographic_proximity": 0.1,
              "reliability_score": 0.2
            }
          },
          "top_n": {
            "type": "integer",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Top N",
            "default": 3
          },
          "page": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Page Size",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "weights"
        ],
        "title": "RecommendationRequest",
        "description": "Request model for MCDA recommendations."
      },
      "RecommendationResult": {
        "properties": {
          "recommended_suppliers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recommended Suppliers",
            "description": "List of recommended suppliers with their scores."
          }
        },
        "type": "object",
        "required": [
          "recommended_suppliers"
        ],
        "title": "RecommendationResult",
        "description": "Data model for a recommendation result."
      },
      "RouteRequest": {
        "properties": {
          "origin": {
            "type": "string",
            "title": "Origin",
            "description": "Origin location name"
          },
          "destination": {
            "type": "string",
            "title": "Destination",
            "description": "Destination location name"
          },
          "weight_tonnes": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Weight Tonnes",
            "description": "Shipment weight in tonnes",
            "default": 1.0
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "description": "Optimization priority: carbon, cost, time, balanced",
            "default": "balanced"
          }
        },
        "type": "object",
        "required": [
          "origin",
          "destination"
        ],
        "title": "RouteRequest",
        "description": "Request model for route optimization"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}